PHPStan to level 7; narrow auth-result shape + PDO false guards
Level-7 surfaces three real categories on top of level 6:

1. AuthService::register() and login() returned the unsafe shape
   array{success: bool, error?: string, user_id?: int}. PHPStan can't
   narrow on a bool flag, so the controller accesses to error / user_id
   tripped the optional-offset check. Switched to a tagged union:
   array{success: true, user_id: int}|array{success: false, error: string}.
   AuthController accesses now narrow cleanly.

2. PDO::query() returns PDOStatement|false. Every call-site chained
   ->fetchAll() / ->fetchColumn() without checking. Added explicit
   false-branch returns in DatabaseSeeder, UserRepository, and
   TrainingService — empty list / zero count, matching the behaviour
   you'd want on an invalid statement.

3. session_name() returns string|false, and setcookie()'s first arg
   is strictly string. Guarded the AuthService::logout() cookie clear.

DebugController::formatBytes had a float index pow into $units —
cast the floor() result to int.

+++ src/Controller/DebugController.php
@@ -122,9 +122,9 @@ class DebugController
     {
         $units = ['B', 'KB', 'MB', 'GB', 'TB'];
         $bytes = max($bytes, 0);
-        $pow = floor(($bytes ? log($bytes) : 0) / log(1024));
+        $pow = (int) floor(($bytes ? log($bytes) : 0) / log(1024));
         $pow = min($pow, count($units) - 1);
-        $bytes /= (1 << (10 * (int)$pow));
+        $bytes /= (1 << (10 * $pow));
         return round($bytes, $precision) . ' ' . $units[$pow];
     }
 }
